backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
litellm==1.57.2
pydantic-settings==2.7.1
pyyaml==6.0.2
sse-starlette==2.2.1
python-dotenv==1.0.1
